# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Ba7ath Forensic API
  version: 3.0.1
  description: >
    API d'authentification et endpoints métiers Ba7ath Forensic (JWT stateless, rate limiting, cache).
servers:
  - url: /api
    description: Local (préfixe API)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    X-RateLimit-Limit:
      description: Nombre max de requêtes autorisées sur la fenêtre en cours
      schema: { type: integer, example: 100 }
    X-RateLimit-Remaining:
      description: Requêtes restantes dans la fenêtre en cours
      schema: { type: integer, example: 87 }
    X-RateLimit-Reset:
      description: Timestamp Unix du reset de la fenêtre
      schema: { type: integer, example: 1694102400 }
    X-RateLimit-Window:
      description: Taille de la fenêtre en secondes
      schema: { type: integer, example: 900 }
    X-RateLimit-Action:
      description: Code logique du limiter appliqué
      schema: { type: string, example: login_attempt }
...
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string, example: "Erreur lors de la connexion" }
        type: { type: string, example: "LOGIN_ERROR" }
        requestId: { type: string, example: "a1b2c3d4e5f6" }
    TokenPair:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
    User:
      type: object
      properties:
        id: { type: string, example: "68be60ce3bb0afd3ce86b8f5" }
        email: { type: string, format: email }
        roles: { type: array, items: { type: string } }
        profile:
          type: object
          properties:
            firstName: { type: string }
            lastName: { type: string }
        security:
          type: object
          properties:
            lastLogin: { type: string, format: date-time }
            lastLoginIP: { type: string }
            lastActivity: { type: string, format: date-time }
            lastActivityIP: { type: string }
            lastActivityUA: { type: string }
    ProfileResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        user: { $ref: '#/components/schemas/User' }
        cached: { type: boolean, example: false }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Connexion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                rememberMe: { type: boolean, default: false }
      responses:
        '200':
          description: Connexion réussie
          headers:
            X-RateLimit-Limit: { $ref: '#/components/headers/X-RateLimit-Limit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/X-RateLimit-Remaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/X-RateLimit-Reset' }
            X-RateLimit-Window: { $ref: '#/components/headers/X-RateLimit-Window' }
            X-RateLimit-Action: { $ref: '#/components/headers/X-RateLimit-Action' }
          content:
            application/json:
              schema:
                type: object
                required: [success, message, user, tokens, session]
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Connexion réussie" }
                  user: { $ref: '#/components/schemas/User' }
                  tokens:
                    type: object
                    properties:
                      accessToken: { type: string }
                      refreshToken: { type: string }
                  session:
                    type: object
                    properties:
                      id: { type: string }
                      expiresAt: { type: string, format: date-time }
                      deviceInfo: { type: object }
        '400':
          description: Données invalides
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Compte suspendu/non vérifié
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '423':
          description: Compte verrouillé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Trop de tentatives
          headers:
            Retry-After:
              description: Secondes avant de réessayer
              schema: { type: integer, example: 900 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Erreur serveur interne
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rafraîchir les tokens JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string, minLength: 32 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  tokens: { $ref: '#/components/schemas/TokenPair' }
                  user: { $ref: '#/components/schemas/User' }
                  requestId: { type: string }
                  timestamp: { type: string, format: date-time }
        '401': { description: Token invalide/révoqué, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: Trop de refresh, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/profile:
    get:
      tags: [auth]
      summary: Profil utilisateur (stateless)
      security: [ { bearerAuth: [] } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ProfileResponse' } } } }
        '401': { description: Non authentifié, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /auth/logout-all:
    post:
      tags: [auth]
      summary: Déconnexion globale de toutes les sessions
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: Sessions révoquées
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  revokedSessions: { type: integer }
                  userId: { type: string }
                  requestId: { type: string }
        '401': { description: Non authentifié, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/change-password:
    post:
      tags: [auth]
      summary: Changement de mot de passe
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword, confirmNewPassword]
              properties:
                currentPassword: { type: string, minLength: 8 }
                newPassword: { type: string, minLength: 8 }
                confirmNewPassword: { type: string, minLength: 8 }
      responses:
        '200': { description: Mot de passe changé, content: { application/json: { schema: { type: object, properties: { success: { type: boolean, example: true }, message: { type: string, example: "Mot de passe modifié avec succès" } } } } } }
        '400': { description: Requête invalide, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Non authentifié, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: Limite horaire atteinte, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/verify-email/{token}:
    get:
      tags: [auth]
      summary: Vérifier l'email à l'aide d'un token
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string, minLength: 32 }
      responses:
        '200':
          description: Email vérifié ou déjà vérifié
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "Email vérifié avec succès" }
                  alreadyVerified: { type: boolean, example: false }
                  verifiedAt: { type: string, format: date-time }
                  requestId: { type: string }
        '400': { description: Token invalide, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '500': { description: Erreur serveur, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /auth/request-password-reset:
    post:
      tags: [auth]
      summary: Demander un reset de mot de passe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Réponse neutre
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  requestId: { type: string }
        '429': { description: Trop de demandes, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/reset-password/{token}:
    post:
      tags: [auth]
      summary: Réinitialiser le mot de passe via token
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string, minLength: 32 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, confirmPassword]
              properties:
                password: { type: string, minLength: 8 }
                confirmPassword: { type: string, minLength: 8 }
      responses:
        '200':
          description: Mot de passe réinitialisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string, example: "Mot de passe réinitialisé" }
                  requestId: { type: string }
        '400': { description: Token invalide/déjà utilisé ou mot de passe faible, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...